+++ services/celery_services/pay_app.py
@@ -7,6 +7,6 @@ from celery import Celery
 from .config import CeleryConfig
 
 
-app = Celery('tasks')
+# app = Celery('tasks')
 
-app.config_from_object(CeleryConfig)
+# app.config_from_object(CeleryConfig)